pyproject.toml
[project]
name = "scraper-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.2",
    "huggingface-hub>=0.23.0",
    "numpy>=2.2.6",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "requests>=2.33.0",
    "sentence-transformers>=2.7.0",
    "transformers>=4.39.3",
    "uvicorn>=0.42.0",
]
